Changes to make device lib compile in linux with new data logger code
* Device lib compiles with data logger code
* Addded changes to make file to fix boost thread issue (same as in engine branch)

Jason




git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti/trunk@1093 8bcac300-4d60-4aec-84ba-b7489f70e69b

+++ device/src/LoggedMeasurement.cpp
@@ -38,7 +38,7 @@ valueIn(valueInput)
 	measurement.setValue(MixedData(0));
 	numberAveragedMeasurements = 0;
 
-	maxNumberToAverage = max(1, saveInterval_secs/measureInterval_secs);
+	maxNumberToAverage = max(1, static_cast<int>(saveInterval_secs/measureInterval_secs));
 
 	resultIsReady = false;
 
@@ -58,7 +58,7 @@ device(sti_device)
 	measurement.setValue(MixedData(0));
 	numberAveragedMeasurements = 0;
 
-	maxNumberToAverage = max(1, saveInterval_secs/measureInterval_secs);
+	maxNumberToAverage = max(1, static_cast<int>(saveInterval_secs/measureInterval_secs));
 
 	resultIsReady = false;
 }
